Trying to go back to the good old days
In some combination of these two files (13 and 13_1) is probably a happy medium. Both now have the red button working (in different places because the white fade and black text fill settings were overriding the red in 13_1). 
The latest problem is that the fade now appears steppy because instead of being drawn every frame, the fade layer is only drawn when text is drawn. Need to break that up somehow to smooth it out.